volume text fix

+++ Assets/Scripts/SaveLoad.cs
@@ -12,9 +12,19 @@ public static class SaveLoad
         SaveLoad.savedGames.Add(Game.current);
         BinaryFormatter bf = new BinaryFormatter();
         //Application.persistentDataPath это строка; выведите ее в логах и вы увидите расположение файла сохранений
-        FileStream file = File.Create(Application.persistentDataPath + "/savedGames.txt");
-        bf.Serialize(file, SaveLoad.savedGames);
-        file.Close();
+        if (!File.Exists(Application.persistentDataPath + "/savedGames.txt"))
+        {
+            FileStream file = File.Create(Application.persistentDataPath + "/savedGames.txt");
+            bf.Serialize(file, SaveLoad.savedGames);
+            file.Close();
+        }
+        else
+        {
+            FileStream file = File.Open(Application.persistentDataPath + "/savedGames.txt", FileMode.Open);
+            bf.Serialize(file, SaveLoad.savedGames);
+            file.Close();
+        }
+
     }
     public static void Load()
     {
